create ExplicitCallToPlusMethod rule This rule detects when the plus(Object) method is called directly in code instead of using the + operator. A groovier way to express this: a.plus(b) is this: a + b
create ExplicitCallToMinusMethod rule
This rule detects when the minus(Object) method is called directly in code instead of using the - operator. A groovier way to express this: a.minus(b) is this: a - b

create ExplicitCallToMultiplyMethod rule
This rule detects when the minus(Object) method is called directly in code instead of using the * operator. A groovier way to express this: a.multiply(b) is this: a * b

create ExplicitCallToDivMethod rule
This rule detects when the div(Object) method is called directly in code instead of using the / operator. A groovier way to express this: a.div(b) is this: a / b

create ExplicitCallToPowerMethod rule
This rule detects when the power(Object) method is called directly in code instead of using the ** operator. A groovier way to express this: a.power(b) is this: a ** b

create ExplicitCallToModMethod rule
This rule detects when the mod(Object) method is called directly in code instead of using the % operator. A groovier way to express this: a.mod(b) is this: a % b


git-svn-id: https://codenarc.svn.sourceforge.net/svnroot/codenarc/trunk@428 58125288-d6d3-43b3-b1ee-2f4d32c6bf4a