create ExplicitCallToOrMethod rule This rule detects when the or(Object) method is called directly in code instead of using the | operator. A groovier way to express this: a.or(b) is this: a | b
create ExplicitCallToAndMethod rule
This rule detects when the and(Object) method is called directly in code instead of using the & operator. A groovier way to express this: a.and(b) is this: a & b

create ExplicitCallToXorMethod rule
This rule detects when the xor(Object) method is called directly in code instead of using the ^ operator. A groovier way to express this: a.xor(b) is this: a ^ b

create ExplicitCallToGetAtMethod rule
This rule detects when the getAt(Object) method is called directly in code instead of using the [] index operator. A groovier way to express this: a.getAt(b) is this: a[b]

create ExplicitCallToLeftShiftMethod rule
This rule detects when the leftShift(Object) method is called directly in code instead of using the << operator. A groovier way to express this: a.leftShift(b) is this: a << b

create ExplicitCallToRightShiftMethod rule
This rule detects when the rightShift(Object) method is called directly in code instead of using the >> operator. A groovier way to express this: a.rightShift(b) is this: a >> b


git-svn-id: https://codenarc.svn.sourceforge.net/svnroot/codenarc/trunk@431 58125288-d6d3-43b3-b1ee-2f4d32c6bf4a