ExplicitCreationOfHashMap This rule checks for the explicit instantiation of a HashMap. In Groovy, it is best to write "new HashMap()" as "[:]", which creates the same object.
ExplicitCreationOfHashSet
This rule checks for the explicit instantiation of a HashSet. In Groovy, it is best to write "new HashSet()" as "[] as Set", which creates the same object.

ExplicitCreationOfTreeSet
This rule checks for the explicit instantiation of a TreeSet. In Groovy, it is best to write "new TreeSet()" as "[] as SortedSet", which creates the same object.

ExplicitCreationOfLinkedList
This rule checks for the explicit instantiation of a LinkedList. In Groovy, it is best to write "new LinkedList()" as "[] as Queue", which creates the same object.

ExplicitCreationOfStack
This rule checks for the explicit instantiation of a Stack. In Groovy, it is best to write "new Stack()" as "[] as Stack", which creates the same object.



git-svn-id: https://codenarc.svn.sourceforge.net/svnroot/codenarc/trunk@434 58125288-d6d3-43b3-b1ee-2f4d32c6bf4a