Update multiline comment checking to shared class
While I was working on updating the `BlankLineBeforePackageRule` to
also exclude blank lines inside of a Multiline comment block, I
noticed the solution was the same as the work I was doing for the
`UnnecessarySemicolonRule`.

I created a new `MultilineCommentChecker` that is used to parse lines
and hold the state of multiline comments.