Allow newline after elvis operator
If there was a newline after an elvis operator, it was throwing a
SpaceAfterOperator warning.

Previously the code was set up to make sure at least one elvis operator
had a space after it. Unfortunately that meant the following line would
have passed:

```groovy
def doubleElvis = fullname ?:lastname ?: 'me'
```

I updated the check if there is instance of an elvis operator with any
non-whitespace character following it.

If you like this method, it could be explored if it is causing problems
with any other operators too.