Update processElvisExpression to be based on columnNumber
I updated this to be based on the columnNumber of the expression.
In the existing test cases, assert statements placed the columnNumber
earlier than the actual elvis operator. I added a comment to describe
why I chose to look for the first index of "?:" starting at the column
index.

I could not come up with a case where the index of the elvis operator
would ever be 0, but in case there was some unknown case, I put a
check to make sure the index had to be greater than 0 for the
preceding check to protect against any index out of bounds problems.
Unfortunately I was not able to create a test for this case.