allow CodeNarc prefix in suppress warnings
in the new version of CodeNarc IntelliJ plugin I register the inspections under `CodeNarc.<rule name>` id to guarantee uniqueness across other plugins. Sadly, thanks to this the suppression action now generates the annotation including the prefix. This behaviour cannot be charged as the method responsible for generating the suppression id is final.