fix: exact path patterns
Fix processing of exact path patterns by matching the pattern against
the file path (relative path), not the absolute path with basedir
prepended (absolute path)

This allows users to specify exact paths within patterns, for example:
-include=subdir/*.groovy
-basedir=/home/myuser

Will now correctly match the file:
/home/myuser/subdir/*.groovy

This changes the behaviour of pattern matches to ignore a forward slash
directly after a double star, for example: **/*.groovy so that it will
match files in the directory or sub directories instead of just sub
directories. This matches the behaviour of ant / globstar.

Also:
* Add bin/ to .gitignore so output files aren't available for commit.
* Add helpful message on assert failure of equal sets.

Fixes #752